Commit Message: "Enhanced cartridge and CPU functionality, improved interrupt handling, and added debugging information"
Added:
- Added "reflect" import in cartridge.go for debugging purposes.
- Introduced RAM_BANK_SIZE constant in cartridge.go.
- Added new fields in Cartridge struct: RamBankSelected, RamBankEnabled, and MemoryModel.
- Initialized RAM banks to maximum size of 128KiB in NewCartridge function.
- Added bankSelectRegister1 and bankSelectRegister2 fields in Mbc1Cartridge struct.
- Implemented SetItem and GetItem methods for Mbc1Cartridge.
- Added debugging information in cpu.go and interrupts.go.

Changed:
- Updated RomBanksCount and RomBankSelected types from uint16 to uint8 in Cartridge struct.
- Modified NewCartridge function to handle changes in Cartridge struct.
- Updated SetItem method in RomOnlyCartridge to handle change in RomBankSelected type.
- Refactored HandleInterrupt and CheckForInterrupts methods in interrupts.go for better interrupt handling.
- Updated GetItem and SetItem methods in motherboard.go to handle changes in Cartridge struct.
- Modified ExecuteInstruction method in cpu.go for better debugging.
- Updated opcodes.go to reflect changes in interrupt handling and debugging.

Removed:
- Removed unnecessary subtraction of address in GetItem and SetItem methods in motherboard.go.

+++ internal/cartridge/rom_only.go
@@ -19,7 +19,7 @@ func (c *RomOnlyCartridge) SetItem(addr uint16, value uint8) {
 		if value == 0 {
 			value = 1
 		}
-		c.parent.RomBankSelected = uint16(value & 0b1)
+		c.parent.RomBankSelected = uint8(value & 0b1)
 		logger.Warnf("Cartridge: Switching to ROM Bank %d\n", c.parent.RomBankSelected)
 	case 0x4000 <= addr && addr < 0xC000:
 		logger.Warn("Cartridge: Writing to RAM is not implemented yet")
